+++ src/structs/result.rs
@@ -9,14 +9,14 @@ pub struct Result {
     pub name: String,
     #[tabled(rename = "Part")]
     pub part: bool,
-    #[tabled(rename = "Output")]
-    pub output: f64,
-    #[tabled(rename = "Duration")]
+    #[tabled(rename = "Answer")]
+    pub output: u128,
+    #[tabled(rename = "Compute time")]
     pub duration: u128,
 }
 
 impl Result {
-    pub fn new(day: usize, name: String, part: bool, output: f64, duration: Duration) -> Self {
+    pub fn new(day: usize, name: String, part: bool, output: u128, duration: Duration) -> Self {
         Self {
             day,
             name,
